refactor: renaming supabase lib files and using supabases getClaims instead of getUser
