+++ src/app/email-password/page.tsx
@@ -1,12 +1,5 @@
 import EmailPasswordDemo from "./EmailPasswordDemo";
-import { createSupabaseServerClient } from "@/lib/supabase/server-client";
 
 export default async function EmailPasswordPage() {
-  const supabase = await createSupabaseServerClient()
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  console.log( { user });
-  return <EmailPasswordDemo user={user} />;
+  return <EmailPasswordDemo />;
 }
